Merges omni-java's _extract_return_type() and target_return_type naming with
the void function support from this branch. Adds fallback to FunctionToOptimize's
return_type attribute when tree-sitter source file parsing is unavailable.

+++ tests/test_languages/test_java/test_instrumentation.py
@@ -226,6 +226,7 @@ public class FibonacciTest {
     }
 }
 """
+
         test_file.write_text(source)
 
         func = FunctionToOptimize(
@@ -2736,7 +2737,7 @@ public class CounterTest__perfinstrumented {
                 }
             }
         }
-        assertEquals(1, _cf_result1_1);
+        assertEquals(1, (int)_cf_result1_1);
     }
 }
 """
